{"schema":"decodex/commit/1","summary":"Split MCP tool schemas from server","authority":"XY-1115"}
